# openapi/fc_openapi.yaml
openapi: 3.0.1
info:
  title: Eclipse XFSC Federated Catalogue
  description: REST API of the XFSC catalogue
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://fc-server.xfsc.org
    description: some future test environment
components:
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: >-
        Forbidden. The user does not have the permission to execute this
        request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: >-
        May contain hints how to solve the error or indicate what went wrong at
        the server. Must not outline any information about the internal
        structure of the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    AdminStats:
      type: object
      properties:
        totalAssets:
          type: integer
          format: int64
          description: Total number of assets in the catalogue
        activeAssets:
          type: integer
          format: int64
          description: Number of active assets
        activeTrustFrameworks:
          type: integer
          format: int64
          description: Number of active trust frameworks
        totalUsers:
          type: integer
          format: int64
          description: Total number of registered users
        totalSchemas:
          type: integer
          format: int64
          description: Total number of schemas
        totalParticipants:
          type: integer
          format: int64
          description: Total number of participants
        graphClaimCount:
          type: integer
          format: int64
          description: Number of claims in the graph database
        graphBackend:
          type: string
          description: The graph database backend type
...
    KeycloakAdminUrl:
      type: object
      properties:
        url:
          type: string
          description: Keycloak Admin Console URL for the catalogue realm
    TrustFrameworkEntry:
      type: object
      properties:
        id:
          type: string
          description: Trust framework identifier
        name:
          type: string
          description: Display name
        serviceUrl:
          type: string
          description: Trust framework service URL
        apiVersion:
          type: string
          description: API version
        timeoutSeconds:
          type: integer
          description: Connection timeout in seconds
        connected:
          type: boolean
          description: Whether the service is currently reachable
        enabled:
          type: boolean
          description: Whether this framework is active
    TrustFrameworkConfigUpdate:
      type: object
      properties:
        serviceUrl:
          type: string
          description: New service URL
        apiVersion:
          type: string
          description: New API version
        timeoutSeconds:
          type: integer
          description: New timeout in seconds
    SchemaValidationStatus:
      type: object
      properties:
        totalSchemaCount:
          type: integer
          format: int64
          description: Total number of schemas across all modules
        modules:
          type: array
          items:
            $ref: '#/components/schemas/SchemaValidationModule'
    SchemaValidationModule:
      type: object
      properties:
        type:
          type: string
          description: Module type (SHACL, JSON_SCHEMA, XML_SCHEMA, OWL)
        name:
          type: string
          description: Display name
        description:
          type: string
          description: Module description
        schemaCount:
          type: integer
          format: int64
          description: Number of schemas for this module
        enabled:
          type: boolean
          description: Whether this module is active
    AssetTypeConfig:
      type: object
      properties:
        enabled:
          type: boolean
          description: Whether asset type restriction is active
        allowedTypes:
          type: array
          items:
            type: string
          description: List of allowed asset types (JSON-LD @type URIs)
    GraphDatabaseStatus:
      type: object
      properties:
        activeBackend:
          type: string
          description: Currently active graph database backend
        connected:
          type: boolean
          description: Whether the graph database is reachable
        claimCount:
          type: integer
          format: int64
          description: Number of claims in the graph database
        version:
          type: string
          description: Graph database version or implementation info
    GraphDatabaseSwitchResult:
      type: object
      properties:
        restartRequired:
          type: boolean
          description: Whether a restart is needed for the switch to take effect
        message:
          type: string
          description: Human-readable status message
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /admin/me:
    get:
      tags:
        - Admin
      summary: Check if the current user has admin access
      operationId: checkAdminAccess
      security:
        - jwt: []
      responses:
        '200':
          description: User has admin access
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/stats:
    get:
      tags:
        - Admin
      summary: Get dashboard statistics
      operationId: getAdminStats
      security:
        - jwt: []
      responses:
        '200':
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /admin/keycloak-url:
    get:
      tags:
        - Admin
      summary: Get the Keycloak Admin Console URL for embedding
      operationId: getKeycloakAdminUrl
      security:
        - jwt: []
      responses:
        '200':
          description: Keycloak Admin Console URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeycloakAdminUrl'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/trust-frameworks:
    get:
      tags:
        - TrustFrameworkAdmin
      summary: List all registered trust frameworks
      operationId: getTrustFrameworks
      security:
        - jwt: []
      responses:
        '200':
          description: List of trust frameworks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrustFrameworkEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/trust-frameworks/{id}/enabled:
    put:
      tags:
        - TrustFrameworkAdmin
      summary: Toggle a trust framework's enabled state
      operationId: setTrustFrameworkEnabled
      security:
        - jwt: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: enabled
          in: query
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: Framework enabled state updated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/trust-frameworks/{id}:
    put:
      tags:
        - TrustFrameworkAdmin
      summary: Update a trust framework's configuration
      operationId: updateTrustFrameworkConfig
      security:
        - jwt: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrustFrameworkConfigUpdate'
      responses:
        '200':
          description: Framework configuration updated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/schema-validation:
    get:
      tags:
        - SchemaValidationAdmin
      summary: Get schema validation module status
      operationId: getSchemaValidationStatus
      security:
        - jwt: []
      responses:
        '200':
          description: Schema validation modules with status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaValidationStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/schema-validation/modules/{type}:
    put:
      tags:
        - SchemaValidationAdmin
      summary: Toggle a schema validation module
      operationId: setSchemaModuleEnabled
      security:
        - jwt: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: enabled
          in: query
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: Module enabled state updated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/asset-types:
    get:
      tags:
        - AssetTypeAdmin
      summary: Get asset type restriction configuration
      operationId: getAssetTypeConfig
      security:
        - jwt: []
      responses:
        '200':
          description: Asset type restriction configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetTypeConfig'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags:
        - AssetTypeAdmin
      summary: Update asset type restriction configuration
      operationId: updateAssetTypeConfig
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetTypeConfig'
      responses:
        '200':
          description: Configuration updated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/asset-types/existing:
    get:
      tags:
        - AssetTypeAdmin
      summary: Get distinct asset types from stored assets
      operationId: getExistingAssetTypes
      security:
        - jwt: []
      responses:
        '200':
          description: List of distinct asset types
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/graph-database:
    get:
      tags:
        - GraphDatabaseAdmin
      summary: Get graph database status
      operationId: getGraphDatabaseStatus
      security:
        - jwt: []
      responses:
        '200':
          description: Graph database status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphDatabaseStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/graph-database/switch:
    post:
      tags:
        - GraphDatabaseAdmin
      summary: Switch the graph database backend (requires restart)
      operationId: switchGraphDatabase
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                backend:
                  type: string
                  description: Target backend (NEO4J, FUSEKI, NONE)
              required:
                - backend
      responses:
        '200':
          description: Switch request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphDatabaseSwitchResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
...